requirements.txt
streamlit
google-genai
python-dotenv
pyyaml
